kuhu
kuhu's page is done, emails and texts and lexia and everything. surprisingly, it was easy to put everything together, go through texts and pictures. i remember for the longest time i couldn't even see her name without feeling really shaken. now im ok, i can even talk to her for a little bit. that's so strange to me, she used to mean so much. the happpiness, the agony, the arguments, the handjobs. it was a lot, and now it doesn't mean anything to me.